App/requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.14
google-generativeai==0.3.2
pyautogui>=0.9.54
pillow>=10.0.0
comtypes; platform_system=="Windows"
